added on target flag to the somatic caller
git-svn-id: https://bsisvn.mayo.edu/main/projects/g/Genome_GPS/trunk@1466 8f329c14-d232-4f17-8684-7ee34322b8dc

+++ concatvcf.sh
@@ -53,8 +53,11 @@ else
 			do
 				if [ $i != ".idx" ]
 				then
-					rm $i
-				fi
+					if [ -s $i ]
+                                        then
+                                            rm $i
+                                        fi
+                                fi
 			done	
         fi
     fi
